+++ apps/backend/src/libs/modules/server-application/server-application.ts
@@ -2,6 +2,7 @@ import { config } from "~/libs/modules/config/config.js";
 import { database } from "~/libs/modules/database/database.js";
 import { logger } from "~/libs/modules/logger/logger.js";
 import { authController } from "~/modules/auth/auth.js";
+import { categoryController } from "~/modules/categories/categories.js";
 import { onboardingController } from "~/modules/onboarding/onboarding.js";
 import { quizController } from "~/modules/quiz/quiz.js";
 import { userController } from "~/modules/users/users.js";
@@ -16,6 +17,7 @@ const apiV1 = new BaseServerApplicationApi(
 	...userController.routes,
 	...quizController.routes,
 	...onboardingController.routes,
+	...categoryController.routes,
 );
 const serverApplication = new BaseServerApplication({
 	apis: [apiV1],