+++ apps/backend/src/modules/categories/category.service.ts
@@ -1,11 +1,15 @@
 import { type Service } from "~/libs/types/types.js";
 
-import { type CategoryEntity } from "./category.entity.js";
+import { CategoryEntity } from "./category.entity.js";
 import { type CategoryRepository } from "./category.repository.js";
 import {
 	type CategoryDto,
+	type CategoryWithScoresDto,
+	type CreateCategoryRequestDto,
+	type GetCategoriesDto,
 	type QuizScoreDto,
 	type QuizScoreRequestDto,
+	type UpdateCategoryRequestDto,
 } from "./libs/types/types.js";
 
 class CategoryService implements Service {
@@ -15,9 +19,17 @@ class CategoryService implements Service {
 		this.categoryRepository = categoryRepository;
 	}
 
-	public convertCategoryEntityToDto(
+	private parseEntityToCategoryDto(
 		categoryEntity: CategoryEntity,
 	): CategoryDto {
+		const { createdAt, id, name, updatedAt } = categoryEntity.toObject();
+
+		return { createdAt, id, name, updatedAt };
+	}
+
+	private parseEntityToCategoryWithScoresDto(
+		categoryEntity: CategoryEntity,
+	): CategoryWithScoresDto {
 		const category = categoryEntity.toObject();
 
 		const scores: QuizScoreDto[] = category.scores.map(
@@ -37,8 +49,12 @@ class CategoryService implements Service {
 		};
 	}
 
-	public create(): Promise<unknown> {
-		return this.categoryRepository.create();
+	public async create(payload: CreateCategoryRequestDto): Promise<CategoryDto> {
+		const categoryEntity = await this.categoryRepository.create(
+			CategoryEntity.initializeNew(payload),
+		);
+
+		return this.parseEntityToCategoryDto(categoryEntity);
 	}
 
 	public async createScore({
@@ -53,8 +69,8 @@ class CategoryService implements Service {
 		});
 	}
 
-	public delete(): Promise<boolean> {
-		return this.categoryRepository.delete();
+	public delete(id: number): Promise<boolean> {
+		return this.categoryRepository.delete(id);
 	}
 
 	public deleteUserScores(userId: number): Promise<number> {
@@ -65,22 +81,37 @@ class CategoryService implements Service {
 		const categoryEntity = await this.categoryRepository.find(id);
 
 		return categoryEntity
-			? this.convertCategoryEntityToDto(categoryEntity)
+			? this.parseEntityToCategoryDto(categoryEntity)
 			: null;
 	}
 
-	public async findAll(): Promise<{ items: CategoryDto[] }> {
+	public async findAll(): Promise<GetCategoriesDto> {
+		const entities = await this.categoryRepository.findAll();
+
+		return {
+			items: entities.map((entity) => this.parseEntityToCategoryDto(entity)),
+		};
+	}
+
+	public async findAllWithScores(): Promise<{
+		items: CategoryWithScoresDto[];
+	}> {
 		const categories = await this.categoryRepository.findAll();
 
 		const items = categories.map((categoryEntity) => {
-			return this.convertCategoryEntityToDto(categoryEntity);
+			return this.parseEntityToCategoryWithScoresDto(categoryEntity);
 		});
 
 		return { items };
 	}
 
-	public update(): Promise<unknown> {
-		return this.categoryRepository.update();
+	public async update(
+		id: number,
+		payload: UpdateCategoryRequestDto,
+	): Promise<CategoryDto> {
+		const categoryEntity = await this.categoryRepository.update(id, payload);
+
+		return this.parseEntityToCategoryDto(categoryEntity);
 	}
 }