+++ apps/backend/src/modules/categories/libs/exceptions/exceptions.ts
@@ -1,1 +1,0 @@
-export { CategoryError } from "shared";