+++ apps/backend/src/modules/categories/libs/types/types.ts
@@ -3,8 +3,9 @@ export { type QuizScoreRequestDto } from "./quiz-score-request-dto.type.js";
 export { type QuizScoreResponseDto } from "./quiz-score-response-dto.type.js";
 export {
 	type CategoryDto,
-	type CategoryRequestDto,
-	type QuizCategoryDto,
-	type QuizGetAllCategoriesResponseDto,
+	type CategoryWithScoresDto,
+	type CreateCategoryRequestDto,
+	type GetCategoriesDto,
 	type QuizScoreDto,
+	type UpdateCategoryRequestDto,
 } from "shared";