+++ apps/backend/src/modules/quiz/quiz.controller.ts
@@ -8,11 +8,6 @@ import { HTTPCode } from "~/libs/modules/http/http.js";
 import { type Logger } from "~/libs/modules/logger/logger.js";
 
 import {
-	type CategoryService,
-	type QuizCategoryDto,
-	type QuizGetAllCategoriesResponseDto,
-} from "../categories/categories.js";
-import {
 	type QuizAnswerService,
 	type QuizAnswersRequestDto,
 } from "../quiz-answers/quiz-answers.js";
@@ -20,31 +15,10 @@ import { type UserDto } from "../users/users.js";
 import { QuizApiPath } from "./libs/enums/enums.js";
 import { quizUserAnswersValidationSchema } from "./libs/validation-schemas/validation-schemas.js";
 
-type Constructor = {
-	categoryService: CategoryService;
-	logger: Logger;
-	quizAnswerService: QuizAnswerService;
-};
-
 /**
  * @swagger
  * components:
  *   schemas:
- *     QuizCategory:
- *       type: object
- *       properties:
- *         id:
- *           type: number
- *           minimum: 1
- *         name:
- *           type: string
- *           example: "Physical"
- *         createdAt:
- *           type: string
- *           format: date-time
- *         updatedAt:
- *           type: string
- *           format: date-time
  *     UserScore:
  *       type: object
  *       properties:
@@ -83,13 +57,11 @@ type Constructor = {
  *           format: int64
  */
 class QuizController extends BaseController {
-	private categoryService: CategoryService;
 	private quizAnswerService: QuizAnswerService;
 
-	constructor({ categoryService, logger, quizAnswerService }: Constructor) {
+	public constructor(logger: Logger, quizAnswerService: QuizAnswerService) {
 		super(logger, APIPath.QUIZ);
 
-		this.categoryService = categoryService;
 		this.quizAnswerService = quizAnswerService;
 
 		this.addRoute({
@@ -105,12 +77,6 @@ class QuizController extends BaseController {
 			validation: {
 				body: quizUserAnswersValidationSchema,
 			},
-		});
-
-		this.addRoute({
-			handler: () => this.getCategories(),
-			method: "GET",
-			path: QuizApiPath.CATEGORIES,
 		});
 	}
 
@@ -164,40 +130,6 @@ class QuizController extends BaseController {
 			}),
 			status: HTTPCode.CREATED,
 		};
-	}
-
-	/**
-	 * @swagger
-	 * /quiz/categories:
-	 *   get:
-	 *     description: Returns an array of quiz categories
-	 *     security:
-	 *       - bearerAuth: []
-	 *     responses:
-	 *       200:
-	 *         description: Successful operation
-	 *         content:
-	 *           application/json:
-	 *             schema:
-	 *               type: object
-	 *               properties:
-	 *                 items:
-	 *                   type: array
-	 *                   items:
-	 *                     $ref: "#/components/schemas/QuizCategory"
-	 */
-	private async getCategories(): Promise<
-		APIHandlerResponse<QuizGetAllCategoriesResponseDto>
-	> {
-		const categories = await this.categoryService.findAll();
-
-		const items: QuizCategoryDto[] = categories.items.map((category) => {
-			const { createdAt, id, name, updatedAt } = category;
-
-			return { createdAt, id, name, updatedAt };
-		});
-
-		return { payload: { items }, status: HTTPCode.OK };
 	}
 }