+++ apps/frontend/src/index.tsx
@@ -5,13 +5,13 @@ import "~/assets/css/styles.css";
 import {
 	App,
 	ProtectedRoute,
-	QuizCategoriesForm,
 	RouterProvider,
 	StoreProvider,
 } from "~/libs/components/components.js";
 import { AppRoute } from "~/libs/enums/enums.js";
 import { store } from "~/libs/modules/store/store.js";
 import { Auth } from "~/pages/auth/auth.jsx";
+import { Chat } from "~/pages/chat/chat.jsx";
 import { NotFound } from "~/pages/not-found/not-found.jsx";
 import { Quiz } from "~/pages/quiz/quiz.jsx";
 
@@ -25,7 +25,7 @@ createRoot(document.querySelector("#root") as HTMLElement).render(
 							{
 								element: (
 									<ProtectedRoute
-										component={<QuizCategoriesForm submitLabel="Retake Quiz" />}
+										component="Root"
 										redirectTo={AppRoute.SIGN_IN}
 									/>
 								),
@@ -40,6 +40,14 @@ createRoot(document.querySelector("#root") as HTMLElement).render(
 								path: AppRoute.SIGN_UP,
 							},
 							{
+								element: <Auth />,
+								path: AppRoute.FORGOT_PASSWORD,
+							},
+							{
+								element: <Auth />,
+								path: AppRoute.RESET_PASSWORD,
+							},
+							{
 								element: (
 									<ProtectedRoute
 										component={<Quiz />}
@@ -47,6 +55,15 @@ createRoot(document.querySelector("#root") as HTMLElement).render(
 									/>
 								),
 								path: AppRoute.QUIZ,
+							},
+							{
+								element: (
+									<ProtectedRoute
+										component={<Chat />}
+										redirectTo={AppRoute.SIGN_IN}
+									/>
+								),
+								path: AppRoute.CHAT,
 							},
 						],
 						element: <App />,