+++ apps/frontend/src/libs/components/components.ts
@@ -7,7 +7,7 @@ export { Link } from "./link/link.js";
 export { Loader } from "./loader/loader.js";
 export { Notification } from "./notification/notification.js";
 export { ProtectedRoute } from "./protected-route/protected-route.js";
-export { QuizCategoriesForm } from "./quiz-categories-selection/quiz-categories-form.js";
+export { QuizCategoriesForm } from "./quiz-categories-form/quiz-categories-form.js";
 export { QuizQuestion } from "./quiz-question/quiz-question.js";
 export { RouterProvider } from "./router-provider/router-provider.js";
 export { Sidebar } from "./sidebar/sidebar.js";