+++ apps/frontend/src/libs/components/input/input.tsx
@@ -15,12 +15,13 @@ type Properties<T extends FieldValues> = {
 	control: Control<T, null>;
 	errors?: FieldErrors<T>;
 	iconName?: IconName;
-	label: string;
+	label?: string;
 	name: FieldPath<T>;
+	onClick?: () => void;
 	onIconClick?: () => void;
 	options?: { label: string; value: string }[];
 	placeholder?: string;
-	type?: "email" | "password" | "radio" | "text";
+	type?: "checkbox" | "email" | "password" | "radio" | "text";
 };
 
 const Input = <T extends FieldValues>({
@@ -29,16 +30,36 @@ const Input = <T extends FieldValues>({
 	iconName,
 	label,
 	name,
+	onClick,
 	onIconClick,
 	options,
 	placeholder = "",
 	type = "text",
 }: Properties<T>): JSX.Element => {
 	const { field } = useFormController({ control, name });
+	const fieldValue = field.value;
 
 	const error = errors?.[name]?.message;
 	const hasError = Boolean(error);
 	const isRadioWithOptions = type === "radio" && options?.length;
+	const isCheckbox = type === "checkbox";
+	const isCheckboxWithOptions = isCheckbox && options?.length;
+
+	const handleCheckboxChange = useCallback(
+		(event: React.ChangeEvent<HTMLInputElement>): void => {
+			const isChecked = event.target.checked;
+			const inputValue = event.target.value;
+
+			if (isChecked) {
+				field.onChange([...fieldValue, inputValue]);
+			} else {
+				field.onChange(
+					(fieldValue as string[]).filter((value) => value !== inputValue),
+				);
+			}
+		},
+		[fieldValue, field],
+	);
 
 	const handleRadioChange = useCallback(
 		(event: React.ChangeEvent<HTMLInputElement>) => {
@@ -46,6 +67,42 @@ const Input = <T extends FieldValues>({
 		},
 		[field],
 	);
+
+	if (isCheckboxWithOptions) {
+		return (
+			<>
+				{options.map((option) => {
+					return (
+						<label className={styles["checkbox-container"]} key={option.value}>
+							<input
+								checked={(fieldValue as string[]).includes(option.value)}
+								onChange={handleCheckboxChange}
+								type={type}
+								value={option.value}
+							/>
+							<span className={styles["checkbox-checkmark"]} />
+							{option.label}
+						</label>
+					);
+				})}
+			</>
+		);
+	}
+
+	if (isCheckbox) {
+		return (
+			<label className={styles["checkbox-container"]}>
+				<input
+					{...field}
+					checked={Boolean(fieldValue)}
+					onClick={onClick}
+					type="checkbox"
+				/>
+				<span className={styles["checkbox-checkmark"]} />
+				{label}
+			</label>
+		);
+	}
 
 	return (
 		<label className={styles["input-wrapper"]}>