+++ apps/frontend/src/libs/enums/enums.ts
@@ -1,3 +1,4 @@
 export { AppRoute } from "./app-route.enum.js";
 export { DataStatus } from "./data-status.enum.js";
+export { NotificationMessage } from "./notification-message.enum.js";
 export { APIPath, AppEnvironment, ContentType, ServerErrorType } from "shared";