+++ apps/frontend/src/libs/hooks/use-app-form/use-app-form.hook.ts
@@ -7,7 +7,6 @@ import {
 	type UseFormGetValues,
 	type UseFormHandleSubmit,
 	type UseFormProps,
-	type UseFormRegister,
 	type UseFormSetError,
 	type UseFormSetValue,
 	type UseFormWatch,
@@ -28,7 +27,6 @@ type ReturnValue<T extends FieldValues = FieldValues> = {
 	errors: FieldErrors<T>;
 	getValues: UseFormGetValues<T>;
 	handleSubmit: UseFormHandleSubmit<T>;
-	register: UseFormRegister<T>;
 	setError: UseFormSetError<T>;
 	setValue: UseFormSetValue<T>;
 	watch: UseFormWatch<T>;
@@ -56,7 +54,6 @@ const useAppForm = <T extends FieldValues = FieldValues>({
 		formState: { errors },
 		getValues,
 		handleSubmit,
-		register,
 		setError,
 		setValue,
 		watch,
@@ -67,7 +64,6 @@ const useAppForm = <T extends FieldValues = FieldValues>({
 		errors,
 		getValues,
 		handleSubmit,
-		register,
 		setError,
 		setValue,
 		watch,