+++ apps/frontend/src/libs/modules/store/store.module.ts
@@ -7,20 +7,26 @@ import { configureStore } from "@reduxjs/toolkit";
 
 import { AppEnvironment } from "~/libs/enums/enums.js";
 import { type Config } from "~/libs/modules/config/config.js";
+import { notification } from "~/libs/modules/notification/notification.js";
+import { storage } from "~/libs/modules/storage/storage.js";
 import { authApi, reducer as authReducer } from "~/modules/auth/auth.js";
-import { quizApi } from "~/modules/quiz/quiz.js";
-import { reducer as quizCategoriesReducer } from "~/modules/quiz-categories/quiz-categories.js";
+import {
+	categoriesApi,
+	reducer as categoriesReducer,
+} from "~/modules/categories/categories.js";
 
 import { handleErrorMiddleware } from "./handle-error.middleware.js";
 
 type RootReducer = {
 	auth: ReturnType<typeof authReducer>;
-	quizCategories: ReturnType<typeof quizCategoriesReducer>;
+	categories: ReturnType<typeof categoriesReducer>;
 };
 
 type ExtraArguments = {
 	authApi: typeof authApi;
-	quizApi: typeof quizApi;
+	categoriesApi: typeof categoriesApi;
+	notification: typeof notification;
+	storage: typeof storage;
 };
 
 class Store {
@@ -44,7 +50,7 @@ class Store {
 			},
 			reducer: {
 				auth: authReducer,
-				quizCategories: quizCategoriesReducer,
+				categories: categoriesReducer,
 			},
 		});
 	}
@@ -52,7 +58,9 @@ class Store {
 	public get extraArguments(): ExtraArguments {
 		return {
 			authApi,
-			quizApi,
+			categoriesApi,
+			notification,
+			storage,
 		};
 	}
 }