+++ apps/frontend/src/modules/categories/categories-api.ts
@@ -1,0 +1,34 @@
+import { APIPath, ContentType } from "~/libs/enums/enums.js";
+import { BaseHTTPApi } from "~/libs/modules/api/api.js";
+import { type HTTP } from "~/libs/modules/http/http.js";
+import { type Storage } from "~/libs/modules/storage/storage.js";
+
+import { CategoriesApiPath } from "./libs/enums/enums.js";
+import { type GetCategoriesDto } from "./libs/types/types.js";
+
+type Constructor = {
+	baseUrl: string;
+	http: HTTP;
+	storage: Storage;
+};
+
+class CategoriesApi extends BaseHTTPApi {
+	public constructor({ baseUrl, http, storage }: Constructor) {
+		super({ baseUrl, http, path: APIPath.CATEGORIES, storage });
+	}
+
+	public async getCategories(): Promise<GetCategoriesDto> {
+		const response = await this.load(
+			this.getFullEndpoint(CategoriesApiPath.ROOT, {}),
+			{
+				contentType: ContentType.JSON,
+				hasAuth: true,
+				method: "GET",
+			},
+		);
+
+		return await response.json<GetCategoriesDto>();
+	}
+}
+
+export { CategoriesApi };