+++ apps/mobile/src/libs/components/planet/libs/maps/size-to-styles.ts
@@ -5,6 +5,7 @@ const sizeToStyles = {
 	md: { borderRadius: 100, height: 100, width: 100 } satisfies ViewStyle,
 	sm: { borderRadius: 86, height: 86, width: 86 } satisfies ViewStyle,
 	xs: { borderRadius: 34, height: 34, width: 34 } satisfies ViewStyle,
+	xxs: { borderRadius: 11, height: 11, width: 11 } satisfies ViewStyle,
 };
 
 export { sizeToStyles };