+++ packages/shared/src/modules/categories/libs/types/create-category-request-dto.type.ts
@@ -1,0 +1,5 @@
+type CreateCategoryRequestDto = {
+	name: string;
+};
+
+export { type CreateCategoryRequestDto };