+++ packages/shared/src/modules/categories/libs/types/get-categories-dto.type.ts
@@ -1,0 +1,5 @@
+import { type CategoryDto } from "./category-dto.type.js";
+
+type GetCategoriesDto = { items: CategoryDto[] };
+
+export { type GetCategoriesDto };