+++ packages/shared/src/modules/quiz/libs/types/quiz-category-dto.type.ts
@@ -1,8 +1,0 @@
-type QuizCategoryDto = {
-	createdAt: string;
-	id: number;
-	name: string;
-	updatedAt: string;
-};
-
-export { type QuizCategoryDto };