+++ packages/shared/src/modules/quiz/quiz.ts
@@ -1,13 +1,9 @@
 export { QuizApiPath } from "./libs/enums/enums.js";
 export {
-	type CategoryDto,
-	type CategoryRequestDto,
 	type QuizAnswerDto,
 	type QuizAnswerRequestDto,
 	type QuizAnswersRequestDto,
 	type QuizAnswersResponseDto,
-	type QuizCategoryDto,
-	type QuizGetAllCategoriesResponseDto,
 	type QuizQuestionDto,
 	type QuizQuestionRequestDto,
 	type QuizScoreDto,